use float

+++ src/game.cpp
@@ -52,7 +52,7 @@ void Game::start() {
     Uint32 time = 0; //time of current frame
     Uint32 oldTime = 0; //time of previous frame
 
-    double frameTime = 0;
+    float frameTime = 0;
 
     onGameCreate();
 
@@ -70,7 +70,7 @@ void Game::start() {
         frame++;
         oldTime = time;
         time = SDL_GetTicks();
-        frameTime = (time - oldTime) / 1000.0; //frameTime is the time this frame has taken, in seconds
+        frameTime = (float) ((time - oldTime) / 1000.0); //frameTime is the time this frame has taken, in seconds
 //        print(1.0 / frameTime); //FPS counter
 //        redraw();
 //        cls();
